+++ CHANGELOG.md
@@ -9,7 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Implement `PartialEq` and `Eq` for `DataValue`.
+- Implement `PartialEq`/`Eq`/`PartialOrd`/`Ord` for `DataValue`.
 
 ## [0.9.4] - 2025-09-01